SELC-1204
* SELC-1204 add to user profile add-to-group button

* SELC-1204 add to user profile add-to-group button

* SELC-1204 add pageRequest to fetchUserGroups

* SELC-1204 fix mockedServices

* SELC-1204 fix addGroup button visibility when user has no group assigned

* SELC-1204 fix control and default option
