+++ src/services/__mocks__/groupsService.ts
@@ -198,6 +198,19 @@ export const mockedGroups: Array<PartyGroupMock> = [
     modifiedAt: new Date('2022-01-01 16:00'),
     modifiedByUserId: 'uid',
   },
+  {
+    id: 'groupId14',
+    name: 'Gruppo14',
+    description: 'Group to have a significant number on prod-io',
+    partyId: 'onboarded',
+    productId: 'prod-io',
+    status: 'ACTIVE',
+    membersIds: ['0'],
+    createdAt: new Date('2022-01-01'),
+    createdByUserId: 'uid',
+    modifiedAt: new Date('2022-01-01 16:00'),
+    modifiedByUserId: 'uid',
+  },
 ].map((o) => ({
   ...o,
   membersCount: o.membersIds.length,
